+++ server/routes/media.js
@@ -2,7 +2,7 @@ const express = require('express');
 const multer = require('multer');
 const imageUpload = require('../utils/image');
 const resumeUpload = require('../utils/resume');
-const avatarUpload = require('../utils/avatar');
+const avatarUpload = require('../utils/aws/avatar');
 const { authenticateToken } = require('../middleware/auth');
 const {
 	badRequestErr,
