+++ server/utils/aws/image.js
